CLEAN UP rewrites and simplifies all back end logic to only the functionality we need
GIT runs into issues when you name multiple files the same thing over the course of different commits

CLEAN UP rewrites and simplifies all back end logic to only the functionality we need
FEAT builds out location checker using Google Maps API

CLEAN UP rewrites and documents all database and data processing logic. Last commit before removing all old logic.

CLEAN UP moves all deprecated server logic to a single folder: server/deprecatedServerLogic

FIX adds csv backups to .gitignore

CLEAN UP grabs useful parts of databaseLogic, all code in deprecatedServerLogic can now be deleted

CLEAN UP cleans up a few things post merge

FIX adds minified version of react and ignores bundle.min.js in client

FEAT builds out location checker using Google Maps API

CLEAN UP rewrites and documents all database and data processing logic. Last commit before removing all old logic.

CLEAN UP moves all deprecated server logic to a single folder: server/deprecatedServerLogic

FIX adds csv backups to .gitignore

CLEAN UP grabs useful parts of databaseLogic, all code in deprecatedServerLogic can now be deleted

CLEAN UP removes deprecatedServerLogic from tracking
