+++ .gitignore
@@ -9,9 +9,10 @@ public/
 .DS_Store
 client/.DS_Store
 client/app/bundle.js
+client/app/bundle.min.js
 client/styles/bundle.css
 client/styles/bundle.min.css
 client/styles/styles.css
 db/*
-client/app/bundle.min.js
-config.js
+server/deprecatedServerLogic/*
+client/app/bundle.min.jsconfig.js
