+++ server/db/sqlSchema/raw_data_import_schema.sql
@@ -1,0 +1,7 @@
+CREATE TABLE 14raw_data_import (
+  activeReposByLang INT(10),
+  repository_language VARCHAR(100),
+  actor_attributes_login VARCHAR(255),
+  actor_attributes_location VARCHAR(255),
+  ID int(10) NOT NULL auto_increment,
+PRIMARY KEY (ID));
