+++ server/deprecatedServerLogic/db/index.js
@@ -2,10 +2,10 @@ var mysql = require('mysql');
 
 dbConnection = mysql.createConnection({
   user: "root",
-  database: "ghdb2"
+  database: "ghdb"
 });
 
-// dbConnection.connect();
+dbConnection.connect();
 
 module.exports = dbConnection;
