+++ server/deprecatedServerLogic/db/sqlSchema/countriesAggregatedSchema.sql
