+++ server/deprecatedServerLogic/db/sqlSchema/placesSchema.sql
