+++ server/deprecatedServerLogic/db/sqlSchema/placesWithGeoSchema.sql
