+++ server/deprecatedServerLogic/db/sqlSchema/rawDataSchema.sql
