+++ server/deprecatedServerLogic/db/sqlSchema/usersAggregatedSchema.sql
