+++ server/deprecatedServerLogic/sqlQueries/aggCountriesToTop10.sql
