+++ server/deprecatedServerLogic/sqlQueries/groupLocations.sql
