+++ server/deprecatedServerLogic/sqlQueries/moveDataToPlaces.sql
